requirements.txt
Flask==2.2.2
pytest~=7.1.3
Werkzeug~=2.2.2
requests~=2.28.0
psutil~=5.9.3
paho-mqtt~=1.6.1
toml~=0.10